Implemented the correct 'self' matching logic according to the spec
Spec: https://w3c.github.io/webappsec-csp/#match-url-to-source-expression

The 'self' matching logic is similar to the host/scheme-source expression
but different enough where it's causing issues because we're just
reusing the csp_source Matches logic.

This is incorrect as causes issues such as the related bug.

This CL covers:
* Added a new MatchesAsSelf function that is used for 'self' expression
matching
* Added unit tests for this function
* Added new test covering the scenario in the bug (wss scheme that matches
scheme)
* Drive-by fixes to some connect-src tests

There is similar work that needs to be done in the content/ CSP
but since I don't want to duplicate any feedback, I will wait until
this patch looks finalized.

Bug: 815142
Change-Id: Ie1d6579e29b9d320e56fcdb556893c7675bd11b8